backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pypdf2==3.0.1
pdfplumber==0.11.0
langchain==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
google-generativeai==0.8.0
textstat==0.7.3
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
rank-bm25==0.2.2
numpy==1.26.4
faiss-cpu==1.9.0.post1
